status, version, fixed a bug with utc date and download folder

+++ cmd/pull_nba.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"fmt"
+	"log"
 
 	"github.com/spf13/cobra"
 	"github.com/stephenhu/stats"
@@ -15,7 +16,7 @@ var (
 		Short: "nba statistics",
 		Long: "nba statistics from the official NBA APIs",
 		Run: func(cmd *cobra.Command, args []string) {
-			pullAll()
+			pullResume()
 		},
 	}
 
@@ -23,26 +24,31 @@ var (
 
 
 func init() {
-
-
-
 } // init
 
 
 func getSchedule() *stats.NbaSchedule {
 
-	schedule := stats.NbaGetSchedule()
+	var schedule = stats.NbaSchedule{}
 
 	fn := fmt.Sprintf("%s/schedule.json", fDir)
+ 
+	if !fileExists(fn) {
 
-	writeJson(schedule, fn)
+		schedule = *stats.NbaGetSchedule()
 
-	return schedule
+		writeJson(schedule, fn)
+
+	} else {
+		readJson(&schedule, fn)
+	}
+
+	return &schedule
 
 } // getSchedule
 
 
-func pullAll() {
+func pullFrom(d string) {
 
 	schedule := getSchedule()
 
@@ -50,28 +56,37 @@ func pullAll() {
 
 		if !stats.IsFutureGame(day.GameDate) {
 
-			for _, game := range day.Games {
-				if game.WeekNumber > 0 {
+			gd := stats.GameDateToString(day.GameDate)
+
+			if gd > d {
+
+				for _, game := range day.Games {
 					
-					box := stats.NbaGetBoxscore(game.ID)
+					if game.WeekNumber > 0 {
+						
+						box := stats.NbaGetBoxscore(game.ID)
+	
+						if len(box.Meta.Time) != 0 && len(box.Meta.Request) != 0 {
+	
+							name := stats.GameDateToString(day.GameDate)
+	
+							// TODO: is this windows friendly?
+							dir := fmt.Sprintf("%s/%s", fDir, name)
+	
+							if !fileExists(dir) {
+								createDir(dir)
+							}
+	
+							fn := fmt.Sprintf("%s/%s.json", dir, game.ID)
 
-					if len(box.Meta.Time) != 0 && len(box.Meta.Request) != 0 {
-
-						name := stats.UtcToFolder(box.Game.GameTime)
-
-						// TODO: is this windows friendly?
-						dir := fmt.Sprintf("%s/%s", fDir, name)
-
-						if !dirExists(dir) {
-							createDir(dir)
+							if !fileExists(fn) {
+								writeJson(box, fn)
+							}
+			
 						}
-
-						fn := fmt.Sprintf("%s/%s.json", dir, game.ID)
-
-						writeJson(box, fn)
-		
+						
 					}
-					
+	
 				}
 
 			}
@@ -80,4 +95,23 @@ func pullAll() {
 
 	}
 
-} // pullAll
+} // pullFrom
+
+
+func pullResume() {
+
+	days := getGameDays()
+
+	log.Println(days)
+	if len(days) == 1 || len(days) == 0 {
+		pullFrom(FROM_SEASON_BEGIN)
+	} else {
+		//log.Println(days[len(days)-1])
+		//log.Println(days[len(days)-2])
+		// this will fail if there other directories with non date names
+		log.Println(days[len(days)-2].Name())
+		pullFrom(days[len(days)-2].Name())
+
+	}
+
+} // pullResume
